Move RenameDialog to interfaces sub-module.
